Executare totala + colorare linii parcurse

+++ structs.cpp
@@ -1,6 +1,6 @@
 #include "structs.h"
 
-map<unsigned short, Linie> liniiDeDesenat;
+map<unsigned, Linie> liniiDeDesenat;
 vector<Arbore> listaArbori;
 map<string, long double> variabile;
 vector<string> listaConsola;
